+++ source/constants/data.py
@@ -1,7 +1,7 @@
 from typing import List, Dict
 
-from source.db.get import getTests, getButtons
-from source.constants.types import TTest, TButtons
+from ..db.get import getTests, getButtons
+from ..constants.types import TTest, TButtons
 
 
 TESTS_CONFIG: List[TTest] = list(getTests())
