+++ source/db/insert.py
@@ -1,22 +1,24 @@
-from source.db.main import get_database
-from source.constants.types import TResult, TTest, TUser, TReminder, TUnfinishedTest
+from pymongo.results import InsertOneResult
+
+from ..db.main import get_database
+from ..constants.types import TResult, TTest, TUser, TReminder, TUnfinishedTest
 
 
-def insertTest(data: TTest):
-    get_database()["tests"].insert_one(data)
+def insertTest(data: TTest) -> InsertOneResult:
+    return get_database()["tests"].insert_one(data)
 
 
-def insertResult(data: TResult):
+def insertResult(data: TResult) -> InsertOneResult:
     return get_database()["results"].insert_one(data)
 
 
-def insertUser(data: TUser):
+def insertUser(data: TUser) -> InsertOneResult:
     return get_database()["users"].insert_one(data).inserted_id
 
 
-def insertReminder(data: TReminder):
-    get_database()["reminders"].insert_one(data)
+def insertReminder(data: TReminder) -> InsertOneResult:
+    return get_database()["reminders"].insert_one(data)
 
 
-def insertUnfinished(data: TUnfinishedTest):
-    get_database()["unfinished_tests"].insert_one(data)
+def insertUnfinished(data: TUnfinishedTest) -> InsertOneResult:
+    return get_database()["unfinished_tests"].insert_one(data)
