+++ source/init/globals.py
@@ -1,5 +1,5 @@
 from typing import List
-from source.constants.types import TGlobals
 
+from ..constants.types import TGlobals
 
 globalsList: List[TGlobals] = []
