+++ source/utils/server/helpers.py
@@ -4,7 +4,7 @@ from dateutil import parser
 from datetime import datetime
 from bson.objectid import ObjectId
 
-from source.utils.helpers import json_serialize
+from ...utils.helpers import json_serialize
 
 def getHandlerFactory(getter, keys=[]):
     def get_handler(request):
